Consolidate URL constants between webapp and docs [S] (#342)
