+++ constants/urls.ts
@@ -1,0 +1,12 @@
+export const URLS = {
+  website: "https://skittles.dev",
+  docs: "https://docs.skittles.dev",
+  github: "https://github.com/chase-manning/skittles",
+  githubExamples: "https://github.com/chase-manning/skittles/tree/main/example",
+  githubReleases: "https://github.com/chase-manning/skittles/releases",
+  githubIssues: "https://github.com/chase-manning/skittles/issues",
+  githubLicense:
+    "https://github.com/chase-manning/skittles/blob/main/LICENSE",
+  npm: "https://www.npmjs.com/package/skittles",
+  npmApi: "https://registry.npmjs.org/skittles/latest",
+};
