+++ docs/constants.ts
@@ -1,5 +1,1 @@
-export const URLS = {
-  website: "https://skittles.dev",
-  github: "https://github.com/chase-manning/skittles",
-  npm: "https://www.npmjs.com/package/skittles",
-};
+export { URLS } from "../constants/urls";
